image-analyzer/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0
transformers
pillow
accelerate
einops
streamlit